Add an option for UTEST_USE_OLD_QPC
This variable controls whether Windows.h is included or not. Including
Windows.h can cause builds to fail, because it contains bugs before
certain versions.

By default UTEST_USE_OLD_QPC is enabled, because Windows 10 LTSC is
based on 17763 and the Windows 10 SDK for that version is the last one
that contained bugs that break builds.